Add dictionary and CSV to the quectelRM500q.py
I've added a dictionary to the code. In this dictionary I store the keys and there values. The keys will be the AT commands and the values are the integer outputs from the AT commands.
Some of the AT commands have multiple values. So I've splitted these into multiple key value pairs using split(',')[].
The next thing I've added is the CSV file. So I want the dictionary to be displayed in a CSV file. The first row schould contain all the keys. And the row below should contain the values. And everytime I run this script the values shoudl stay and the new values should be added and the keys won't if the keys are already in the header.
Warning: this code isn't working. Every time I run this code I get disconnected from the testbed. I've created a new file 'testdict.py' to test if dictionaries work. This file works fine. But in my quectelRM500q.py file it doens't work.
The other troubleshoot I've done is delete the CSV code and just use the dictionry and print it. Even then it crashes and disconnects. So there is a bug in dictionary.