Fixed the disconnect problem.
I solved the problem. I did this by not placing the result of the def 'exectute command' directly in the directory and splitting it into Several parts using .'split(',').
This allows us to view our output as different parts. Finally, there was only a problem with the first output value of the key.
It still had the at command followed by a ':'. So we applied an extra split to the initial values to get just the value.