Add new script for AT commands network setup
I've added a new script that is used to setup the UE with AT commands. This script can have an input from the argparse.
It's called 'setup' If it's true it will just enable and disable arplane mode (we use this to connect is the setup is already done), on the other hand if it's false you will do the whole setup of the UE.
I've done this to separate 'the setup and enable/diables airplaine mode)' from the capture of the data.
Beceause we don't need to setup the while UE every time or enable/disable airplane everytime we want to capture data. It's often done in the beginning, not every time.