fine fune iperf2 script
I've corrected the errors in the iperf3 script. So I've added a dollar sign before arguments that were missing one and I've deleted the first check for all the arguments.
I did this because when you want to you tcp instead of udp you don't need to write something. I my previous code this would give a problem, because it checked for all the arguments.