Update my scripts
To start off I've changed my quectelRM500q.py script to make sure that writing from the dictionary to the csv files keeps going.
Secondly I've added a new python script to extract more information from the iperf script. It will extract data from JSON and organize it in a dictionary.
This includes measurement,date & time, transfer information, bitrate, jitter, lost packets, lost percentage, and protocol.
Italso adds 1 second to the date & time of the measurement for each interval in the data, using the timedelta function from the datetime module.
The original date and time of the measurement is extracted from the start.timestamp.time field in the JSON data, and for each interval, the loop adds one second using timedelta(seconds=i), where i is the loop index.
Afterwards the code write the data to a CSV file. It will create one if there isn't one.t will create one if there isn't one, and appending to it if it already exists.
If it's appending it's only appending the values and not the keys.