Add Interactive map 5G indoor network testing
I've visualized my data using plotly.  I have created a scatter plot on X an Y coord from a dataframe. This scatter is on top of an image (beacon floor plan in 1280x720 pixels). 
If you hover over the measurement points you can see the data that was measured on that exact place.